refactoring, curl scripts added
