+++ src/Shared.Core/Services/VirtualDesktops/Interfaces/IVirtualDesktopService.cs
@@ -1,4 +1,4 @@
-﻿namespace AMTools.Shared.Core.Services.Interfaces
+﻿namespace AMTools.Shared.Core.Services.VirtualDesktops.Interfaces
 {
     public interface IVirtualDesktopService
     {
