+++ src/Shared.Core/Services/VirtualDesktops/Interfaces/IVirtualDesktopVersionService.cs
@@ -1,4 +1,4 @@
-﻿namespace AMTools.Shared.Core.Services.Interfaces
+﻿namespace AMTools.Shared.Core.Services.VirtualDesktops.Interfaces
 {
     public interface IVirtualDesktopVersionService
     {
