+++ src/Shared.Core/Services/VirtualDesktops/VirtualDesktopService.cs
@@ -6,8 +6,9 @@ using System.Threading.Tasks;
 using AMTools.Core.Services.Logging;
 using AMTools.Shared.Core.Models;
 using AMTools.Shared.Core.Services.Interfaces;
+using AMTools.Shared.Core.Services.VirtualDesktops.Interfaces;
 
-namespace AMTools.Shared.Core.Services
+namespace AMTools.Shared.Core.Services.VirtualDesktops
 {
     public class VirtualDesktopService : IVirtualDesktopService
     {
