+++ src/Shared.Core/Services/VirtualDesktops/VirtualDesktopVersionService.cs
@@ -1,9 +1,9 @@
 ﻿using System;
 using System.IO;
 using AMTools.Core.Services.Logging;
-using AMTools.Shared.Core.Services.Interfaces;
+using AMTools.Shared.Core.Services.VirtualDesktops.Interfaces;
 
-namespace AMTools.Shared.Core.Services
+namespace AMTools.Shared.Core.Services.VirtualDesktops
 {
     public class VirtualDesktopVersionService : IVirtualDesktopVersionService
     {
