+++ src/Web/_Skripte/Desktop umschalten.cmd
@@ -1,0 +1,14 @@
+@echo off
+call Skriptkonfiguration.cmd
+
+echo Es wird zum nächsten Desktop gewechselt.
+
+curl -X POST -H "accept: text/plain" -H "Content-Length: 0" "%SERVERADRESSE%/api/Desktop/Switch"
+echo.
+
+set arg1=%1
+IF "%arg1%" == "nopause" GOTO END
+
+pause
+
+:END
