+++ src/Web/_Skripte/Skriptkonfiguration.cmd
@@ -1,0 +1,3 @@
+REM Achtung: Die Serveradresse darf nicht mit / enden
+
+set SERVERADRESSE=http://localhost:5000
